Changed Top 250 weighting .5 -> .6.
With a weighting of 0.5, a game with 97% and 121818 votes was placed
above a game with 97% 204909 votes. This seems strictly incorrect, as
both have a similar approval rating but the game with almost double the
number of votes is placed second.

For reference, the database snapeshots were 109 and 110 and the games
were Euro Truck Simulator 2 and Left 4 Dead 2.