Changed Top 250 weighting .6 -> .7.
A game with 97% approval and 64K votes was sitting above a game with 97%
approval and 122K votes. Since it has double the votes it should not be
in such close contention. This change swaps their positions and also
forces a few more well known titles into the list and flushes out a few
obscurities.

For reference, the database snapshots were 120 and 119 and the games
were Mount & Blade: Warband and Borderlands 2.