Changed Bottom 100 weight from Bayesian 500 -> 4000. Added "secret" Buyer's Remorse list (linked from Bottom 100).
The bottom 100 has been completely redesigned to focus far less on vote
volume and more on (dis)approval rate. The number of votes needed to
enter this list has been reduced to around 500 votes, the same threshold
needed for Steam to rank a game as overwhelmingly positive or negative.

Insodoing, larger commercial disappointments have mostly been flushed out
of this list, so a separate Buyers Remorse list was added to feature
these.