Changed Build Monitor to show overdue message instead of ready on page load.
Due to recent changes in the way the next build date is determined, if
the build was overdue it was flashing the "ready" message easter egg.
Previously, the scheduled cron would always run before the timer expired
so this state was never encountered, but now we rely on the import job
to complete successfully, which may not happen in case of an error. On
page load, if the next build date has elapsed, we now show an overdue
message instead of the easter egg, which is supposed to stay hidden!