Changed clock to remove itself instead of showing overdue message.
If the page build is overdue this usually indicates an error occurred
during the scheduled import, in which case the clock won't be correct
for at least the next 24 hours, when the next scheduled import begins.
Better to remove the clock than display an animated message incorrectly
promising an imminent build.