Changed video footer to no longer be a flow element.
Previously the video footer contributed to the overall size of the video
container. By removing it from the flow the video is positioned dead
vertical center, and in case the viewport is equal to the video size,
there will no longer be a space for the footer to expand into. This also
has the effect that hovering over the space where the footer _would_ be
does not cause it to appear.

It is thought it is preferable for a hidden element to not contribute to
the layout of neighbouring video elements. However it is my personal
preference that the video be slightly negatively offset in the vertical
axes as previously. It is also unclear whether it really is beneficial
to avoid cutting off part of the top of the video in lieu of the footer
navivigation bar. Such small screen resolutions are undesirable in
general but also unavoidable on mobile devices so it is a tough problem
to solve.