Added refined hover states for both checkbox modes.
Previously we were lazy and just applied a filter to the entire element
tree, which looked tacky in the on/off mode and bad in the A/B mode. In
A/B mode we now do not apply any hover effect to the deselected text.
For both modes, we do not modify the track's background colour, which
"grounds" it in the page better.