Allowed VrTop250Ranking to be empty due to upstream data errors.
On 2022/12/07 Steam stopped showing VR platforms (probably temporarily).
This allows the ranking to be empty and generalizes the behaviour in the
new AllowEmptyRanking trait. Previously, this specific exception was
only granted to MostPlayedRanking but has now been extended to VR and
the front-end notice has been migrated from most_played.twig to the
general ranking.twig base class.