Assign correct height to ranking images even when not loaded.
This is important to avoid layout shift as images are lazy loaded during
scrolling, and to ensure the "bottom" link actually takes the user to
the bottom of the page consistently, even when images at the bottom
haven't loaded yet. This is achieved using aspect ratio rather than
height, since width is calculated dynamically for small (mobile)
resolutions and thus height would be too, if not for the convenience of
the aspect ratio feature.